apps/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "narpisa-pdf-worker"
version = "0.2.0"
description = "FastAPI worker for transient PDF fetching and parsing."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "google-auth",
  "google-cloud-tasks",
  "httpx",
  "pydantic-settings",
  "pypdf",
  "requests",
  "supabase",
  "uvicorn[standard]",
]

[project.optional-dependencies]
dev = [
  "black",
  "mypy",
  "pytest",
  "pytest-asyncio",
  "ruff",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
exclude = ["tests"]
